+++ Assets/Data/Item/Item Data.asset
@@ -17,7 +17,7 @@ MonoBehaviour:
   _itemImage: {fileID: 0}
   _description: 
   _damage: 0
-  _range: 10
+  _range: 1
   _attackSpeed: 0
   _additionalHp: 0
   _additionalStemina: 0
